% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_x.R
\name{bin_x2}
\alias{bin_x2}
\title{bin_x2}
\usage{
bin_x2(x, x_min = 60, x_max = 1000, pixel = 224)
}
\arguments{
\item{x}{x}

\item{x_min}{x_min}

\item{x_max}{x_max}

\item{pixel}{pixel}
}
\value{
a new vector
}
\description{
bin_x2
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
