% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mz_shift.R
\name{get_mz_shift}
\alias{get_mz_shift}
\title{get_mz_shift}
\usage{
get_mz_shift(
  mz = c(87.04406, 90.05499, 94.06512),
  mean = 1.12361,
  sd = 4.076444
)
}
\arguments{
\item{mz}{mz}

\item{mean}{mean}

\item{sd}{sd}
}
\value{
a new vector
}
\description{
get_mz_shift
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
