% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_conflicts.R
\name{pseudomsir_conflicts}
\alias{pseudomsir_conflicts}
\title{Conflicts between the pseudomsir and other packages}
\usage{
pseudomsir_conflicts()
}
\value{
pseudomsir_conflicts
}
\description{
This function lists all the conflicts between packages in the pseudomsir
and other packages that you have loaded.
}
\details{
There are four conflicts that are deliberately ignored: \code{intersect},
\code{union}, \code{setequal}, and \code{setdiff}
from dplyr. These functions
make the base equivalents generic, so shouldn't negatively affect any
existing code.
}
\examples{
pseudomsir_conflicts()
}
