% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_pseudomsir_logo.R
\name{pseudomsir_logo}
\alias{pseudomsir_logo}
\title{pseudomsir_logo}
\usage{
pseudomsir_logo()
}
\value{
logo
}
\description{
pseudomsir_logo
}
\examples{
pseudomsir_logo()
}
\author{
Xiaotao Shen
\email{shenxt@stanford.edu}
}
