% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_utils.R
\name{pseudomsir_packages}
\alias{pseudomsir_packages}
\title{List all packages in the pseudomsir}
\usage{
pseudomsir_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include pseudomsir in the list?}
}
\value{
pseudomsir_packages
}
\description{
List all packages in the pseudomsir
}
